push

+++ app/Views/salesnyamitra/lihat_sales.php
@@ -43,9 +43,19 @@
                                 <td><?php echo $d["alamat"] ?></td>
                                 <td><?php echo $d["jenis_kelamin"] ?></td>
                                 <td><?php echo $d["email"] ?></td>
-                                <td>
-                                        <a href="<?php echo base_url() ?>/datasalmit/hapus_sales/<?php echo $d["email"] ?> "><button class="btn btn-danger btn-sm rounded-0" type="button" data-toggle="tooltip" data-placement="top" title="Delete"><i class="fa fa-trash"> Hapus</i></button>
-                                </td>
+                                <?php
+                                    if ($d["status_kepegawaian"] == "non pegawai") { ?>
+                                        <td>
+                                            <a href="<?php echo base_url() ?>/datasalmit/terima_pegawai/<?php echo $d["email"] ?> "><button class="btn btn-warning btn-sm rounded-0" type="button" data-toggle="tooltip" data-placement="top" title="Edit"><i class="fas fa-check"></i>Terima</button>
+                                            <a href="<?php echo base_url() ?>/datasalmit/hapus_sales/<?php echo $d["email"] ?> "><button class="btn btn-danger btn-sm rounded-0" type="button" data-toggle="tooltip" data-placement="top" title="Delete"><i class="fa fa-trash"> Hapus</i></button>
+                                        </td>
+                                    <?php } ?>
+                                    <?php
+                                    if ($d["status_kepegawaian"] == "pegawai") { ?>
+                                        <td>
+                                            <a href="<?php echo base_url() ?>/datasalmit/hapus_sales/<?php echo $d["email"] ?> "><button class="btn btn-danger btn-sm rounded-0" type="button" data-toggle="tooltip" data-placement="top" title="Delete"><i class="fa fa-trash"> Hapus</i></button>
+                                        </td>
+                                    <?php } ?>
                             </tr>
                         <?php } ?>
                         </tr>
